refactor: 데미지 타입에 대한 면역 태그 매핑 설정을 프로젝트 설정으로 이전
