+++ Config/DefaultGASGameFramework.ini
@@ -4,3 +4,6 @@ DefinitionDataTableMap=(("/Script/CoreUObject.Class'/Script/GGFCharacterSystem.G
 [/Script/GGFItemSystem.GGFItemSystemSetting]
 ItemDataTable=/Game/AutoWorld/Data/DT_Item.DT_Item
 ItemConfigDataTableMap=(("/Script/CoreUObject.Class'/Script/GGFItemSystem.GGFItemConfig_Pickup'", "/Game/AutoWorld/Data/DT_Pickup.DT_Pickup"))
+
+[/Script/GASExtension.GASExtensionSetting]
+DamageImmunityMap=(((Tag=(TagName="Damage.Type.Test")), (Tag=(TagName="Buff.Immunity.Test"))),((Tag=(TagName="Damage.Type.Physical")), (Tag=(TagName="Buff.Immunity.Physical"))),((Tag=(TagName="Damage.Type.Fire")), (Tag=(TagName="Buff.Immunity.Fire"))),((Tag=(TagName="Damage.Type.Water")), (Tag=(TagName="Buff.Immunity.Water"))),((Tag=(TagName="Damage.Type.Electricity")), (Tag=(TagName="Buff.Immunity.Electricity"))),((Tag=(TagName="Damage.Type.Ice")), (Tag=(TagName="Buff.Immunity.Ice"))))
