+++ Source/GASExtension/Private/GASExtensionSetting.cpp
@@ -1,0 +1,3 @@
+﻿// Fill out your copyright notice in the Description page of Project Settings.
+
+#include "GASExtensionSetting.h"
