+++ Source/GASGameFramework/Public/GASGameFramework.h
@@ -17,6 +17,9 @@ private:
     // GAS를 위한 설정
     void InitializeForGameplayAbilitySystem();
 
+    // 설정 파일 리로드
+    void ReloadConfigs();
+
 #if WITH_EDITOR
     // 디테일 패널에 커스텀 섹션 추가
     void RegisterSectionMappings();
